style: move StatsSheet to sheets package
