+++ app/src/main/java/com/github/libretube/ui/sheets/StatsSheet.kt
@@ -1,4 +1,4 @@
-package com.github.libretube.ui.dialogs
+package com.github.libretube.ui.sheets
 
 import android.annotation.SuppressLint
 import android.os.Bundle
